requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.0.0
httpx==0.24.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
jinja2==3.1.2
xhtml2pdf==0.2.11 